feat: 반려동물 등록 페이지에 세부 정보 입력 필드 및 훅 추가
- petDetailFields.tsx 파일 생성하여 반려동물 세부 정보 입력 필드 정의
- usePetDetail.ts 파일 생성하여 반려동물 세부 정보 관리 로직 구현
- PetDetailPage.tsx에서 기존 로직을 usePetDetail 훅으로 리팩토링하여 코드 간결화 및 재사용성 향상
